Setup initial rust buildenv docker images
Four different docker images are built. They follow the official Rust
docker images for both the stable and the nightly toolchains, with the
regular and the slim versions being provided.

Each installation step is provided as a different install-*.sh script
that is copied into the image, executed and then removed (for minimizing
the image size). Some packages also need to be installed in the system
due to dependencies, in which case the apt cache is cleaned up
afterwards, also for minimizing the final image size.

Besides the tools added as part of the build environment, only one
script is left in the image (echo-buildenv.sh). This script just outputs
a summary of the build environment state. It is automatically executed
when the image is launched so that anyone can get a quick glance at the
exact versions of each tool provided in the build environment. It is
also executed as the last step of building the image itself for a quick
summary of the image contents.