Add CircleCI configuration
CircleCI is used for building the different docker images and pushing
them to Docker Hub (as the experience with docker's build infrastructure
is less than ideal). Each branch will build the images and save them to
a tarball but only the master branch will push the images. Note that in
order to push the images a docker-credentials context with the
DOCKER_PASSWORD environment variable must be set up in CircleCI.