Add nightly and weekly builds
Docker Hub's automated builds trigger whenever the base images are
updated. In order to ensure that the CircleCI-driven images also keep up
with base image updates two schedules are added: a daily schedule that
should pick up nightly tags with a minimum delay and a weekly schedule
to pick up the stable tags.