Add cargo-cache installation
The cargo-cache tool allows for querying the cache state as well as
performing a number of different maintenance tasks. One of them is
automatically performed at the end of the image build: removing all
unnecessary directories from the cache. This slims down the size of the
final images as those caches were only used for building the different
tools.