Add cargo-build-deps to the image
This tool can be very useful when caching dependencies in either Docker
image layers or a CI system. However, the original tool present in
crates.io seems to be unmaintained (as in, doesn't seem to work with a
project that was created recently). Therefore a fork of the tool is
installed instead.